# new spectral initialization option: a truncated SVD-warmed lobpcg 