@lmcinnes 2023-03-29
Thanks for this. I have in the past run up against the spectral layout having odd issues, and had as much (or slightly more, since I never came up with a solution like this) difficulty as you. This makes a lot of sense as far as I can tell. I may actually end up adding a deprecation notice and swapping this in as the standard spectral layout in due course. Thanks for all the work figuring this out.